CIRC-1589 Update tests
